+++ tests/mixins.py
@@ -153,7 +153,7 @@ class InteractionMixin(BaseMixin):
     #       here, but everything I've tried seems to run afoul of pyright type checker.
     #       So I'm just using `...` here for now, even though it sucks.
     async def run(self, command: Command[CogT, CogCallbackP, None], **kwargs: Any) -> None:
-        interaction = kwargs.get("interaction", None)
+        interaction = kwargs.get("interaction")
         if not interaction:
             kwargs["interaction"] = self.interaction
         callback = command.callback