+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v11.5.0](https://github.com/lexicalunit/spellbot/releases/tag/v11.5.0) - 2024-10-16
+
 ### Changed
 
 - Updated python dependencies.